+++ src/main/java/com/realitysink/cover/nodes/expression/CoverInverseCommaUnsignedIntNode.java
@@ -22,7 +22,6 @@ import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.nodes.NodeInfo;
 import com.realitysink.cover.nodes.CoverType;
 import com.realitysink.cover.nodes.CoverTypedExpressionNode;
-import com.realitysink.cover.nodes.INT32;
 
 /**
  * The "comma" operator. Evaluates both expressions and returns the value of the second.
@@ -33,7 +32,7 @@ import com.realitysink.cover.nodes.INT32;
 @NodeChildren({@NodeChild("leftNode"), @NodeChild("rightNode")})
 public abstract class CoverInverseCommaUnsignedIntNode extends CoverTypedExpressionNode {
     @Specialization
-    protected INT32 add(INT32 left, Object right) { // FIXME: are there faster ways to ignore the first node?
+    protected long add(long left, Object right) {
         return left;
     }
     public CoverType getType() {
