+++ src/main/java/com/realitysink/cover/nodes/local/CoverReadSignedIntArrayValueNode.java
@@ -21,13 +21,12 @@ import com.oracle.truffle.api.dsl.NodeChildren;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.realitysink.cover.nodes.CoverType;
 import com.realitysink.cover.nodes.CoverTypedExpressionNode;
-import com.realitysink.cover.nodes.INT32;
 
 @NodeChildren({@NodeChild("array"),@NodeChild("expressionNode")})
 public abstract class CoverReadSignedIntArrayValueNode extends CoverTypedExpressionNode {
     @Specialization
-    public INT32 readSignedInt(INT32[] array, long index) {
-        return array[(int) index];
+    public long readSignedInt(long[] array, long index) {
+        return (int)array[(int) index];
     }
     
     @Override
